implemented security, bug fixes, data adjustments

+++ api/src/main/java/com/fundmate/api/service/impl/AccountServiceImpl.java
@@ -9,6 +9,7 @@ import com.fundmate.api.repository.AccountRepository;
 import com.fundmate.api.repository.UserRepository;
 import com.fundmate.api.service.AccountService;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -28,21 +29,22 @@ public class AccountServiceImpl implements AccountService {
         this.accountMapper = accountMapper;
     }
 
-    @Override
-    public AccountResponse createAccount(AccountRequest accountRequest, Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+    private User getCurrentUser() {
+        return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    }
 
+    @Override
+    public AccountResponse createAccount(AccountRequest accountRequest) {
         Account account = accountMapper.toEntity(accountRequest);
-        account.setUser(user);
+        account.setUser(getCurrentUser());
         Account savedAccount = accountRepository.save(account);
 
         return accountMapper.toResponse(savedAccount);
     }
 
     @Override
-    public List<AccountResponse> getAccountsByUserId(Long userId) {
-        return accountRepository.findByUserId(userId).stream()
+    public List<AccountResponse> getAccounts() {
+        return accountRepository.findByUserId(getCurrentUser().getId()).stream()
                 .map(accountMapper::toResponse)
                 .collect(Collectors.toList());
     }
@@ -51,6 +53,11 @@ public class AccountServiceImpl implements AccountService {
     public AccountResponse getAccountById(Long accountId) {
         Account account = accountRepository.findById(accountId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found"));
+
+        if (!account.getUser().getId().equals(getCurrentUser().getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        }
+
         return accountMapper.toResponse(account);
     }
 
@@ -58,6 +65,11 @@ public class AccountServiceImpl implements AccountService {
     public AccountResponse updateAccount(Long accountId, AccountRequest accountRequest) {
         Account account = accountRepository.findById(accountId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found"));
+
+        if (!account.getUser().getId().equals(getCurrentUser().getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        }
+
         account.setAccountName(accountRequest.getAccountName());
         account.setBalance(accountRequest.getBalance());
 
@@ -67,6 +79,13 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public void deleteAccount(Long accountId) {
+        Account account = accountRepository.findById(accountId)
+                        .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Account not found"));
+
+        if (!account.getUser().getId().equals(getCurrentUser().getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Access denied");
+        }
+
         accountRepository.deleteById(accountId);
     }
 }
